Code Cleanups.

+++ app/src/main/java/com/pr0gramm/app/services/VoteService.java
@@ -203,7 +203,7 @@ public class VoteService {
     }
 
     public Observable<Map<Long, Vote>> getTagVotes(List<Tag> tags) {
-        List<Long> ids = transform(tags, tag -> (long) tag.getId());
+        List<Long> ids = transform(tags, tag -> tag.getId());
         return findCachedVotes(CachedVote.Type.TAG, ids);
     }
